create Project-SafephoneReadme.md
The purpose of this project is to collaborate with users to find a solution for ensuring security for your cell-phone.
The Idea is that every time you restart/reboot your phone Project-Safephone will run to ensure that unwanted 
software, malware, trackers, or any other type of malicious code can enable a remote user to hijack your phone, 
data, or identity. Collaborators welcome as I am new to coding and need all of the help I can get. This project
is going to start out Open-Source provided the purpose of the project can be fulfilled and still remain Open and free
to the public.